fake corpus files
the 'splitcorpus.py' takes 'dummy.txt' as input corpus and creates 'mug'
and 'test' as ouput. I upload them anyway as an example.
The cooler way to do it would be to ask the user for the input file and
the names for the resulting corpora files, but I won't have time to
change that till Wednesday.
The whole thing took me about 3h, I guess. It doesn't look good, but it
works.